Implemented all functionality of products and dishes tabs. Code refactoring in js code and also code refactoring in java code.

+++ src/main/java/com/kostenko/pp/data/entity/Product.java
@@ -21,7 +21,7 @@ public class Product {
     @Column(name = "name", nullable = false)
     private String name;
     @Column(name = "energy", nullable = false)
-    private int energy;
+    private Double energy;
     @Column(name = "typeId", nullable = false)
     private Long typeId;
 }
